store both raw and coordinated omission metrics
