+++ src/lib.rs
@@ -438,9 +438,7 @@ pub mod util;
 mod worker;
 
 use chrono::prelude::*;
-use chrono::Duration;
 use gumdrop::Options;
-use itertools::Itertools;
 use lazy_static::lazy_static;
 #[cfg(feature = "gaggle")]
 use nng::Socket;
@@ -449,7 +447,7 @@ use rand::thread_rng;
 use serde::{Deserialize, Serialize};
 use simplelog::*;
 use std::collections::hash_map::DefaultHasher;
-use std::collections::{BTreeMap, HashMap};
+use std::collections::BTreeMap;
 use std::hash::{Hash, Hasher};
 use std::path::PathBuf;
 use std::sync::{
@@ -3392,329 +3390,6 @@ impl GooseAttack {
         Ok(())
     }
 
-    // Write an HTML-formatted report, if enabled.
-    async fn write_html_report(
-        &mut self,
-        goose_attack_run_state: &mut GooseAttackRunState,
-    ) -> Result<(), GooseError> {
-        // Only write the report if enabled.
-        if let Some(report_file) = goose_attack_run_state.report_file.as_mut() {
-            // Prepare report summary variables.
-            let started = self.metrics.started.unwrap();
-            let start_time = started.format("%Y-%m-%d %H:%M:%S").to_string();
-            let end_time = (started + Duration::seconds(self.metrics.duration as i64))
-                .format("%Y-%m-%d %H:%M:%S")
-                .to_string();
-            let host = match self.get_configuration_host() {
-                Some(h) => h.to_string(),
-                None => "".to_string(),
-            };
-
-            // Prepare requests and responses variables.
-            let mut request_metrics = Vec::new();
-            let mut response_metrics = Vec::new();
-            let mut aggregate_total_count = 0;
-            let mut aggregate_fail_count = 0;
-            let mut aggregate_response_time_counter: usize = 0;
-            let mut aggregate_response_time_minimum: usize = 0;
-            let mut aggregate_response_time_maximum: usize = 0;
-            let mut aggregate_response_times: BTreeMap<usize, usize> = BTreeMap::new();
-            for (request_key, request) in self.metrics.requests.iter().sorted() {
-                let method = format!("{}", request.method);
-                // The request_key is "{method} {name}", so by stripping the "{method} "
-                // prefix we get the name.
-                // @TODO: consider storing the name as a field in GooseRequestMetricAggregate.
-                let name = request_key
-                    .strip_prefix(&format!("{} ", request.method))
-                    .unwrap()
-                    .to_string();
-                let total_request_count = request.success_count + request.fail_count;
-                let (requests_per_second, failures_per_second) = metrics::per_second_calculations(
-                    self.metrics.duration,
-                    total_request_count,
-                    request.fail_count,
-                );
-                // Prepare per-request metrics.
-                request_metrics.push(report::RequestMetric {
-                    method: method.to_string(),
-                    name: name.to_string(),
-                    number_of_requests: total_request_count,
-                    number_of_failures: request.fail_count,
-                    response_time_average: format!(
-                        "{:.2}",
-                        request.total_response_time as f32 / request.response_time_counter as f32
-                    ),
-                    response_time_minimum: request.min_response_time,
-                    response_time_maximum: request.max_response_time,
-                    requests_per_second: format!("{:.2}", requests_per_second),
-                    failures_per_second: format!("{:.2}", failures_per_second),
-                });
-
-                // Prepare per-response metrics.
-                response_metrics.push(report::get_response_metric(
-                    &method,
-                    &name,
-                    &request.response_times,
-                    request.response_time_counter,
-                    request.min_response_time,
-                    request.max_response_time,
-                ));
-
-                // Collect aggregated request and response metrics.
-                aggregate_total_count += total_request_count;
-                aggregate_fail_count += request.fail_count;
-                aggregate_response_time_counter += request.total_response_time;
-                aggregate_response_time_minimum = metrics::update_min_time(
-                    aggregate_response_time_minimum,
-                    request.min_response_time,
-                );
-                aggregate_response_time_maximum = metrics::update_max_time(
-                    aggregate_response_time_maximum,
-                    request.max_response_time,
-                );
-                aggregate_response_times =
-                    metrics::merge_times(aggregate_response_times, request.response_times.clone());
-            }
-
-            // Prepare aggregate per-request metrics.
-            let (aggregate_requests_per_second, aggregate_failures_per_second) =
-                metrics::per_second_calculations(
-                    self.metrics.duration,
-                    aggregate_total_count,
-                    aggregate_fail_count,
-                );
-            request_metrics.push(report::RequestMetric {
-                method: "".to_string(),
-                name: "Aggregated".to_string(),
-                number_of_requests: aggregate_total_count,
-                number_of_failures: aggregate_fail_count,
-                response_time_average: format!(
-                    "{:.2}",
-                    aggregate_response_time_counter as f32 / aggregate_total_count as f32
-                ),
-                response_time_minimum: aggregate_response_time_minimum,
-                response_time_maximum: aggregate_response_time_maximum,
-                requests_per_second: format!("{:.2}", aggregate_requests_per_second),
-                failures_per_second: format!("{:.2}", aggregate_failures_per_second),
-            });
-
-            // Prepare aggregate per-response metrics.
-            response_metrics.push(report::get_response_metric(
-                "",
-                "Aggregated",
-                &aggregate_response_times,
-                aggregate_total_count,
-                aggregate_response_time_minimum,
-                aggregate_response_time_maximum,
-            ));
-
-            // Compile the request metrics template.
-            let mut requests_rows = Vec::new();
-            for metric in request_metrics {
-                requests_rows.push(report::request_metrics_row(metric));
-            }
-
-            // Compile the response metrics template.
-            let mut responses_rows = Vec::new();
-            for metric in response_metrics {
-                responses_rows.push(report::response_metrics_row(metric));
-            }
-
-            // Only build the tasks template if --no-task-metrics isn't enabled.
-            let tasks_template: String;
-            if !self.configuration.no_task_metrics {
-                let mut task_metrics = Vec::new();
-                let mut aggregate_total_count = 0;
-                let mut aggregate_fail_count = 0;
-                let mut aggregate_task_time_counter: usize = 0;
-                let mut aggregate_task_time_minimum: usize = 0;
-                let mut aggregate_task_time_maximum: usize = 0;
-                let mut aggregate_task_times: BTreeMap<usize, usize> = BTreeMap::new();
-                for (task_set_counter, task_set) in self.metrics.tasks.iter().enumerate() {
-                    for (task_counter, task) in task_set.iter().enumerate() {
-                        if task_counter == 0 {
-                            // Only the taskset_name is used for task sets.
-                            task_metrics.push(report::TaskMetric {
-                                is_task_set: true,
-                                task: "".to_string(),
-                                name: task.taskset_name.to_string(),
-                                number_of_requests: 0,
-                                number_of_failures: 0,
-                                response_time_average: "".to_string(),
-                                response_time_minimum: 0,
-                                response_time_maximum: 0,
-                                requests_per_second: "".to_string(),
-                                failures_per_second: "".to_string(),
-                            });
-                        }
-                        let total_run_count = task.success_count + task.fail_count;
-                        let (requests_per_second, failures_per_second) =
-                            metrics::per_second_calculations(
-                                self.metrics.duration,
-                                total_run_count,
-                                task.fail_count,
-                            );
-                        let average = match task.counter {
-                            0 => 0.00,
-                            _ => task.total_time as f32 / task.counter as f32,
-                        };
-                        task_metrics.push(report::TaskMetric {
-                            is_task_set: false,
-                            task: format!("{}.{}", task_set_counter, task_counter),
-                            name: task.task_name.to_string(),
-                            number_of_requests: total_run_count,
-                            number_of_failures: task.fail_count,
-                            response_time_average: format!("{:.2}", average),
-                            response_time_minimum: task.min_time,
-                            response_time_maximum: task.max_time,
-                            requests_per_second: format!("{:.2}", requests_per_second),
-                            failures_per_second: format!("{:.2}", failures_per_second),
-                        });
-
-                        aggregate_total_count += total_run_count;
-                        aggregate_fail_count += task.fail_count;
-                        aggregate_task_times =
-                            metrics::merge_times(aggregate_task_times, task.times.clone());
-                        aggregate_task_time_counter += &task.counter;
-                        aggregate_task_time_minimum =
-                            metrics::update_min_time(aggregate_task_time_minimum, task.min_time);
-                        aggregate_task_time_maximum =
-                            metrics::update_max_time(aggregate_task_time_maximum, task.max_time);
-                    }
-                }
-
-                let (aggregate_requests_per_second, aggregate_failures_per_second) =
-                    metrics::per_second_calculations(
-                        self.metrics.duration,
-                        aggregate_total_count,
-                        aggregate_fail_count,
-                    );
-                task_metrics.push(report::TaskMetric {
-                    is_task_set: false,
-                    task: "".to_string(),
-                    name: "Aggregated".to_string(),
-                    number_of_requests: aggregate_total_count,
-                    number_of_failures: aggregate_fail_count,
-                    response_time_average: format!(
-                        "{:.2}",
-                        aggregate_response_time_counter as f32 / aggregate_total_count as f32
-                    ),
-                    response_time_minimum: aggregate_task_time_minimum,
-                    response_time_maximum: aggregate_task_time_maximum,
-                    requests_per_second: format!("{:.2}", aggregate_requests_per_second),
-                    failures_per_second: format!("{:.2}", aggregate_failures_per_second),
-                });
-                let mut tasks_rows = Vec::new();
-                // Compile the task metrics template.
-                for metric in task_metrics {
-                    tasks_rows.push(report::task_metrics_row(metric));
-                }
-
-                tasks_template = report::task_metrics_template(&tasks_rows.join("\n"));
-            } else {
-                tasks_template = "".to_string();
-            }
-
-            // Only build the tasks template if --no-task-metrics isn't enabled.
-            let errors_template: String;
-            if !self.metrics.errors.is_empty() {
-                let mut error_rows = Vec::new();
-                for error in self.metrics.errors.values() {
-                    error_rows.push(report::error_row(error));
-                }
-                errors_template = report::errors_template(&error_rows.join("\n"));
-            } else {
-                errors_template = "".to_string();
-            }
-
-            // Only build the status_code template if --status-codes is enabled.
-            let status_code_template: String;
-            if self.configuration.status_codes {
-                let mut status_code_metrics = Vec::new();
-                let mut aggregated_status_code_counts: HashMap<u16, usize> = HashMap::new();
-                for (request_key, request) in self.metrics.requests.iter().sorted() {
-                    let method = format!("{}", request.method);
-                    // The request_key is "{method} {name}", so by stripping the "{method} "
-                    // prefix we get the name.
-                    // @TODO: consider storing the name as a field in GooseRequestMetricAggregate.
-                    let name = request_key
-                        .strip_prefix(&format!("{} ", request.method))
-                        .unwrap()
-                        .to_string();
-
-                    // Build a list of status codes, and update the aggregate record.
-                    let codes = metrics::prepare_status_codes(
-                        &request.status_code_counts,
-                        &mut Some(&mut aggregated_status_code_counts),
-                    );
-
-                    // Add a row of data for the status code table.
-                    status_code_metrics.push(report::StatusCodeMetric {
-                        method,
-                        name,
-                        status_codes: codes,
-                    });
-                }
-
-                // Build a list of aggregate status codes.
-                let aggregated_codes =
-                    metrics::prepare_status_codes(&aggregated_status_code_counts, &mut None);
-
-                // Add a final row of aggregate data for the status code table.
-                status_code_metrics.push(report::StatusCodeMetric {
-                    method: "".to_string(),
-                    name: "Aggregated".to_string(),
-                    status_codes: aggregated_codes,
-                });
-
-                // Compile the status_code metrics rows.
-                let mut status_code_rows = Vec::new();
-                for metric in status_code_metrics {
-                    status_code_rows.push(report::status_code_metrics_row(metric));
-                }
-
-                // Compile the status_code metrics template.
-                status_code_template =
-                    report::status_code_metrics_template(&status_code_rows.join("\n"));
-            } else {
-                // If --status-codes is not enabled, return an empty template.
-                status_code_template = "".to_string();
-            }
-
-            // Compile the report template.
-            let report = report::build_report(
-                &start_time,
-                &end_time,
-                &host,
-                report::GooseReportTemplates {
-                    requests_template: &requests_rows.join("\n"),
-                    responses_template: &responses_rows.join("\n"),
-                    tasks_template: &tasks_template,
-                    status_codes_template: &status_code_template,
-                    errors_template: &errors_template,
-                },
-            );
-
-            // Write the report to file.
-            if let Err(e) = report_file.write(report.as_ref()).await {
-                return Err(GooseError::InvalidOption {
-                    option: "--report-file".to_string(),
-                    value: self.get_report_file_path().unwrap(),
-                    detail: format!("Failed to create report file: {}", e),
-                });
-            };
-            // Be sure the file flushes to disk.
-            report_file.flush().await?;
-
-            info!(
-                "wrote html report file to: {}",
-                self.get_report_file_path().unwrap()
-            );
-        }
-
-        Ok(())
-    }
-
     // Reset the GooseAttackRunState before starting a load test. This is to allow a Controller
     // to stop and start the load test multiple times, for example from a UI.
     async fn reset_run_state(
@@ -3856,7 +3531,7 @@ impl GooseAttack {
                     // Collect all metrics sent by GooseUser threads.
                     self.sync_metrics(&mut goose_attack_run_state, true).await?;
                     // Write an html report, if enabled.
-                    self.write_html_report(&mut goose_attack_run_state).await?;
+                    //self.write_html_report(&mut goose_attack_run_state).await?;
                     // Shutdown Goose or go into an idle waiting state.
                     if goose_attack_run_state.shutdown_after_stop {
                         self.set_attack_phase(&mut goose_attack_run_state, AttackPhase::Shutdown);
