+++ src/metrics.rs
@@ -346,31 +346,10 @@ pub struct GooseRequestMetricAggregate {
     ///
     /// For example: [`GooseMethod::Get`].
     pub method: GooseMethod,
-    /// Per-response-time counters, tracking how often pages are returned with this response time.
-    ///
-    /// All response times between 1 and 100ms are stored without any rounding. Response times between
-    /// 100 and 500ms are rounded to the nearest 10ms and then stored. Response times betwee 500 and
-    /// 1000ms are rounded to the nearest 100ms. Response times larger than 1000ms are rounded to the
-    /// nearest 1000ms.
-    pub response_times: BTreeMap<usize, usize>,
-    /// The shortest response time seen so far.
-    ///
-    /// For example a `min_response_time` of `3` means the quickest response for this method-path
-    /// pair returned in 3 milliseconds. This value is not rounded.
-    pub min_response_time: usize,
-    /// The longest response time seen so far.
-    ///
-    /// For example a `max_response_time` of `2013` means the slowest response for this method-path
-    /// pair returned in 2013 milliseconds. This value is not rounded.
-    pub max_response_time: usize,
-    /// Total combined response times seen so far.
-    ///
-    /// A running total of all response times returned for this method-path pair.
-    pub total_response_time: usize,
-    /// Total number of response times seen so far.
-    ///
-    /// A count of how many requests have been tracked for this method-path pair.
-    pub response_time_counter: usize,
+    /// The raw data seen from actual requests.
+    pub raw_data: GooseRequestMetricTimingData,
+    /// Combines the raw data with statistically generated Coordinated Omission Metrics.
+    pub coordinated_omission_data: Option<GooseRequestMetricTimingData>,
     /// Per-status-code counters, tracking how often each response code was returned for this request.
     pub status_code_counts: HashMap<u16, usize>,
     /// Total number of times this path-method request resulted in a successful (2xx) status code.
@@ -394,11 +373,8 @@ impl GooseRequestMetricAggregate {
         GooseRequestMetricAggregate {
             path: path.to_string(),
             method,
-            response_times: BTreeMap::new(),
-            min_response_time: 0,
-            max_response_time: 0,
-            total_response_time: 0,
-            response_time_counter: 0,
+            raw_data: GooseRequestMetricTimingData::new(None),
+            coordinated_omission_data: None,
             status_code_counts: HashMap::new(),
             success_count: 0,
             fail_count: 0,
@@ -406,62 +382,27 @@ impl GooseRequestMetricAggregate {
         }
     }
 
-    /// Track response time.
-    pub(crate) fn set_response_time(&mut self, response_time: u64) {
-        // Perform this conversin only once, then re-use throughout this funciton.
-        let response_time_usize = response_time as usize;
-
-        // Update minimum if this one is fastest yet.
-        if self.min_response_time == 0
-            || (response_time_usize > 0 && response_time_usize < self.min_response_time)
-        {
-            self.min_response_time = response_time_usize;
+    pub(crate) fn record_time(&mut self, time_elapsed: u64, coordinated_omission_mitigation: bool) {
+        // Only add time_elapsed to raw_data if the time wasn't generated by Coordinated
+        // Omission Mitigation.
+        if !coordinated_omission_mitigation {
+            self.raw_data.record_time(time_elapsed);
         }
 
-        // Update maximum if this one is slowest yet.
-        if response_time_usize > self.max_response_time {
-            self.max_response_time = response_time_usize;
+        // A Coordinated Omission data object already exists, add a new time into the data.
+        if let Some(coordinated_omission_data) = self.coordinated_omission_data.as_mut() {
+            coordinated_omission_data.record_time(time_elapsed);
         }
-
-        // Update total_response time, adding in this one.
-        self.total_response_time += response_time_usize;
-
-        // Each time we store a new response time, increment counter by one.
-        self.response_time_counter += 1;
-
-        // Round the response time so we can combine similar times together and
-        // minimize required memory to store and push upstream to the parent.
-        // No rounding for 1-100ms response times.
-        let rounded_response_time = if response_time < 100 {
-            response_time_usize
-        }
-        // Round to nearest 10 for 100-500ms response times.
-        else if response_time < 500 {
-            ((response_time as f64 / 10.0).round() * 10.0) as usize
-        }
-        // Round to nearest 100 for 500-1000ms response times.
-        else if response_time < 1000 {
-            ((response_time as f64 / 100.0).round() * 100.0) as usize
-        }
-        // Round to nearest 1000 for all larger response times.
+        // Create a new Coordinated Omission data object by cloning the raw data.
         else {
-            ((response_time as f64 / 1000.0).round() * 1000.0) as usize
-        };
-
-        let counter = match self.response_times.get(&rounded_response_time) {
-            // We've seen this response_time before, increment counter.
-            Some(c) => {
-                debug!("got {:?} counter: {}", rounded_response_time, c);
-                *c + 1
+            let mut coordinated_omission_data = self.raw_data.clone();
+            // If this time_elapsed was generated by Coordinated Omission Mitigation, it doesn't
+            // exist in the raw_data, so add it.
+            if coordinated_omission_mitigation {
+                coordinated_omission_data.record_time(time_elapsed);
             }
-            // First time we've seen this response time, initialize counter.
-            None => {
-                debug!("no match for counter: {}", rounded_response_time);
-                1
-            }
-        };
-        self.response_times.insert(rounded_response_time, counter);
-        debug!("incremented {} counter: {}", rounded_response_time, counter);
+            self.coordinated_omission_data = Some(coordinated_omission_data);
+        }
     }
 
     /// Increment counter for status code, creating new counter if first time seeing status code.
@@ -492,6 +433,110 @@ impl Ord for GooseRequestMetricAggregate {
 impl PartialOrd for GooseRequestMetricAggregate {
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
         Some(self.cmp(other))
+    }
+}
+/// Collects per-request timing metrics.
+#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize)]
+pub struct GooseRequestMetricTimingData {
+    /// Per-response-time counters, tracking how often pages are returned with this response time.
+    ///
+    /// All response times between 1 and 100ms are stored without any rounding. Response times between
+    /// 100 and 500ms are rounded to the nearest 10ms and then stored. Response times betwee 500 and
+    /// 1000ms are rounded to the nearest 100ms. Response times larger than 1000ms are rounded to the
+    /// nearest 1000ms.
+    pub times: BTreeMap<usize, usize>,
+    /// The shortest response time seen so far.
+    ///
+    /// For example a `min_response_time` of `3` means the quickest response for this method-path
+    /// pair returned in 3 milliseconds. This value is not rounded.
+    pub minimum_time: usize,
+    /// The longest response time seen so far.
+    ///
+    /// For example a `max_response_time` of `2013` means the slowest response for this method-path
+    /// pair returned in 2013 milliseconds. This value is not rounded.
+    pub maximum_time: usize,
+    /// Total combined response times seen so far.
+    ///
+    /// A running total of all response times returned for this method-path pair.
+    pub total_time: usize,
+    /// Total number of response times seen so far.
+    ///
+    /// A count of how many requests have been tracked for this method-path pair.
+    pub counter: usize,
+}
+impl GooseRequestMetricTimingData {
+    /// Create a new GooseRequestMetricAggregate object.
+    pub(crate) fn new(metric_data: Option<&GooseRequestMetricTimingData>) -> Self {
+        trace!("new GooseRequestMetricTimingData");
+        // Simply clone the exiting metric_data.
+        if let Some(data) = metric_data {
+            data.clone()
+        // Create a new empty metric_data.
+        } else {
+            GooseRequestMetricTimingData {
+                times: BTreeMap::new(),
+                minimum_time: 0,
+                maximum_time: 0,
+                total_time: 0,
+                counter: 0,
+            }
+        }
+    }
+
+    /// Record a new time.
+    pub(crate) fn record_time(&mut self, time_elapsed: u64) {
+        // Perform this conversin only once, then re-use throughout this funciton.
+        let time = time_elapsed as usize;
+
+        // Update minimum if this one is fastest yet.
+        if time > 0 || (self.minimum_time == 0 || time < self.minimum_time) {
+            self.minimum_time = time;
+        }
+
+        // Update maximum if this one is slowest yet.
+        if time > self.maximum_time {
+            self.maximum_time = time;
+        }
+
+        // Update total time, adding in this one.
+        self.total_time += time;
+
+        // Each time we store a new time, increment counter by one.
+        self.counter += 1;
+
+        // Round the time so we can combine similar times together and
+        // minimize required memory to store and push upstream to the parent.
+        // No rounding for 1-100ms times.
+        let rounded_time = if time_elapsed < 100 {
+            time
+        }
+        // Round to nearest 10 for 100-500ms times.
+        else if time_elapsed < 500 {
+            ((time_elapsed as f64 / 10.0).round() * 10.0) as usize
+        }
+        // Round to nearest 100 for 500-1000ms times.
+        else if time_elapsed < 1000 {
+            ((time_elapsed as f64 / 100.0).round() * 100.0) as usize
+        }
+        // Round to nearest 1000 for all larger times.
+        else {
+            ((time_elapsed as f64 / 1000.0).round() * 1000.0) as usize
+        };
+
+        let counter = match self.times.get(&rounded_time) {
+            // We've seen this elapsed time before, increment counter.
+            Some(c) => {
+                debug!("got {:?} counter: {}", rounded_time, c);
+                *c + 1
+            }
+            // First time we've seen this elapsed time, initialize counter.
+            None => {
+                debug!("no match for counter: {}", rounded_time);
+                1
+            }
+        };
+        debug!("incremented {} counter: {}", rounded_time, counter);
+        self.times.insert(rounded_time, counter);
     }
 }
 
@@ -1290,6 +1335,7 @@ impl GooseMetrics {
             fmt,
             " ------------------------------------------------------------------------------"
         )?;
+        /*
         for (request_key, request) in self.requests.iter().sorted() {
             let average = match request.response_time_counter {
                 0 => 0.0,
@@ -1331,6 +1377,7 @@ impl GooseMetrics {
                 avg_precision = average_precision,
             )?;
         }
+        */
         if self.requests.len() > 1 {
             let average = match aggregate_response_time_counter {
                 0 => 0.0,
@@ -1398,6 +1445,7 @@ impl GooseMetrics {
             fmt,
             " ------------------------------------------------------------------------------"
         )?;
+        /*
         for (request_key, request) in self.requests.iter().sorted() {
             // Iterate over user response times, and merge into global response times.
             aggregate_response_times =
@@ -1465,6 +1513,7 @@ impl GooseMetrics {
                 ),
             )?;
         }
+        */
         if self.requests.len() > 1 {
             writeln!(
                 fmt,
@@ -1845,7 +1894,7 @@ impl GooseAttack {
         }
         // Store a new metric.
         else {
-            merge_request.set_response_time(request_metric.response_time);
+            merge_request.record_time(request_metric.response_time, request_metric.coordinated_omission_elapsed > 0);
             if self.configuration.status_codes {
                 merge_request.set_status_code(request_metric.status_code);
             }
@@ -2011,6 +2060,330 @@ impl GooseAttack {
             self.metrics.duration = 0;
         }
     }
+
+    /*
+    // Write an HTML-formatted report, if enabled.
+    pub(crate) async fn write_html_report(
+        &mut self,
+        goose_attack_run_state: &mut GooseAttackRunState,
+    ) -> Result<(), GooseError> {
+        // Only write the report if enabled.
+        if let Some(report_file) = goose_attack_run_state.report_file.as_mut() {
+            // Prepare report summary variables.
+            let started = self.metrics.started.unwrap();
+            let start_time = started.format("%Y-%m-%d %H:%M:%S").to_string();
+            let end_time = (started + Duration::seconds(self.metrics.duration as i64))
+                .format("%Y-%m-%d %H:%M:%S")
+                .to_string();
+            let host = match self.get_configuration_host() {
+                Some(h) => h.to_string(),
+                None => "".to_string(),
+            };
+
+            // Prepare requests and responses variables.
+            let mut request_metrics = Vec::new();
+            let mut response_metrics = Vec::new();
+            let mut aggregate_total_count = 0;
+            let mut aggregate_fail_count = 0;
+            let mut aggregate_response_time_counter: usize = 0;
+            let mut aggregate_response_time_minimum: usize = 0;
+            let mut aggregate_response_time_maximum: usize = 0;
+            let mut aggregate_response_times: BTreeMap<usize, usize> = BTreeMap::new();
+            for (request_key, request) in self.metrics.requests.iter().sorted() {
+                let method = format!("{}", request.method);
+                // The request_key is "{method} {name}", so by stripping the "{method} "
+                // prefix we get the name.
+                let name = request_key
+                    .strip_prefix(&format!("{} ", request.method))
+                    .unwrap()
+                    .to_string();
+                let total_request_count = request.success_count + request.fail_count;
+                let (requests_per_second, failures_per_second) = metrics::per_second_calculations(
+                    self.metrics.duration,
+                    total_request_count,
+                    request.fail_count,
+                );
+                // Prepare per-request metrics.
+                request_metrics.push(report::RequestMetric {
+                    method: method.to_string(),
+                    name: name.to_string(),
+                    number_of_requests: total_request_count,
+                    number_of_failures: request.fail_count,
+                    response_time_average: format!(
+                        "{:.2}",
+                        request.total_response_time as f32 / request.response_time_counter as f32
+                    ),
+                    response_time_minimum: request.min_response_time,
+                    response_time_maximum: request.max_response_time,
+                    requests_per_second: format!("{:.2}", requests_per_second),
+                    failures_per_second: format!("{:.2}", failures_per_second),
+                });
+
+                // Prepare per-response metrics.
+                response_metrics.push(report::get_response_metric(
+                    &method,
+                    &name,
+                    &request.response_times,
+                    request.response_time_counter,
+                    request.min_response_time,
+                    request.max_response_time,
+                ));
+
+                // Collect aggregated request and response metrics.
+                aggregate_total_count += total_request_count;
+                aggregate_fail_count += request.fail_count;
+                aggregate_response_time_counter += request.total_response_time;
+                aggregate_response_time_minimum = metrics::update_min_time(
+                    aggregate_response_time_minimum,
+                    request.min_response_time,
+                );
+                aggregate_response_time_maximum = metrics::update_max_time(
+                    aggregate_response_time_maximum,
+                    request.max_response_time,
+                );
+                aggregate_response_times =
+                    metrics::merge_times(aggregate_response_times, request.response_times.clone());
+            }
+
+            // Prepare aggregate per-request metrics.
+            let (aggregate_requests_per_second, aggregate_failures_per_second) =
+                metrics::per_second_calculations(
+                    self.metrics.duration,
+                    aggregate_total_count,
+                    aggregate_fail_count,
+                );
+            request_metrics.push(report::RequestMetric {
+                method: "".to_string(),
+                name: "Aggregated".to_string(),
+                number_of_requests: aggregate_total_count,
+                number_of_failures: aggregate_fail_count,
+                response_time_average: format!(
+                    "{:.2}",
+                    aggregate_response_time_counter as f32 / aggregate_total_count as f32
+                ),
+                response_time_minimum: aggregate_response_time_minimum,
+                response_time_maximum: aggregate_response_time_maximum,
+                requests_per_second: format!("{:.2}", aggregate_requests_per_second),
+                failures_per_second: format!("{:.2}", aggregate_failures_per_second),
+            });
+
+            // Prepare aggregate per-response metrics.
+            response_metrics.push(report::get_response_metric(
+                "",
+                "Aggregated",
+                &aggregate_response_times,
+                aggregate_total_count,
+                aggregate_response_time_minimum,
+                aggregate_response_time_maximum,
+            ));
+
+            // Compile the request metrics template.
+            let mut requests_rows = Vec::new();
+            for metric in request_metrics {
+                requests_rows.push(report::request_metrics_row(metric));
+            }
+
+            // Compile the response metrics template.
+            let mut responses_rows = Vec::new();
+            for metric in response_metrics {
+                responses_rows.push(report::response_metrics_row(metric));
+            }
+
+            // Only build the tasks template if --no-task-metrics isn't enabled.
+            let tasks_template: String;
+            if !self.configuration.no_task_metrics {
+                let mut task_metrics = Vec::new();
+                let mut aggregate_total_count = 0;
+                let mut aggregate_fail_count = 0;
+                let mut aggregate_task_time_counter: usize = 0;
+                let mut aggregate_task_time_minimum: usize = 0;
+                let mut aggregate_task_time_maximum: usize = 0;
+                let mut aggregate_task_times: BTreeMap<usize, usize> = BTreeMap::new();
+                for (task_set_counter, task_set) in self.metrics.tasks.iter().enumerate() {
+                    for (task_counter, task) in task_set.iter().enumerate() {
+                        if task_counter == 0 {
+                            // Only the taskset_name is used for task sets.
+                            task_metrics.push(report::TaskMetric {
+                                is_task_set: true,
+                                task: "".to_string(),
+                                name: task.taskset_name.to_string(),
+                                number_of_requests: 0,
+                                number_of_failures: 0,
+                                response_time_average: "".to_string(),
+                                response_time_minimum: 0,
+                                response_time_maximum: 0,
+                                requests_per_second: "".to_string(),
+                                failures_per_second: "".to_string(),
+                            });
+                        }
+                        let total_run_count = task.success_count + task.fail_count;
+                        let (requests_per_second, failures_per_second) =
+                            metrics::per_second_calculations(
+                                self.metrics.duration,
+                                total_run_count,
+                                task.fail_count,
+                            );
+                        let average = match task.counter {
+                            0 => 0.00,
+                            _ => task.total_time as f32 / task.counter as f32,
+                        };
+                        task_metrics.push(report::TaskMetric {
+                            is_task_set: false,
+                            task: format!("{}.{}", task_set_counter, task_counter),
+                            name: task.task_name.to_string(),
+                            number_of_requests: total_run_count,
+                            number_of_failures: task.fail_count,
+                            response_time_average: format!("{:.2}", average),
+                            response_time_minimum: task.min_time,
+                            response_time_maximum: task.max_time,
+                            requests_per_second: format!("{:.2}", requests_per_second),
+                            failures_per_second: format!("{:.2}", failures_per_second),
+                        });
+
+                        aggregate_total_count += total_run_count;
+                        aggregate_fail_count += task.fail_count;
+                        aggregate_task_times =
+                            metrics::merge_times(aggregate_task_times, task.times.clone());
+                        aggregate_task_time_counter += &task.counter;
+                        aggregate_task_time_minimum =
+                            metrics::update_min_time(aggregate_task_time_minimum, task.min_time);
+                        aggregate_task_time_maximum =
+                            metrics::update_max_time(aggregate_task_time_maximum, task.max_time);
+                    }
+                }
+
+                let (aggregate_requests_per_second, aggregate_failures_per_second) =
+                    metrics::per_second_calculations(
+                        self.metrics.duration,
+                        aggregate_total_count,
+                        aggregate_fail_count,
+                    );
+                task_metrics.push(report::TaskMetric {
+                    is_task_set: false,
+                    task: "".to_string(),
+                    name: "Aggregated".to_string(),
+                    number_of_requests: aggregate_total_count,
+                    number_of_failures: aggregate_fail_count,
+                    response_time_average: format!(
+                        "{:.2}",
+                        aggregate_response_time_counter as f32 / aggregate_total_count as f32
+                    ),
+                    response_time_minimum: aggregate_task_time_minimum,
+                    response_time_maximum: aggregate_task_time_maximum,
+                    requests_per_second: format!("{:.2}", aggregate_requests_per_second),
+                    failures_per_second: format!("{:.2}", aggregate_failures_per_second),
+                });
+                let mut tasks_rows = Vec::new();
+                // Compile the task metrics template.
+                for metric in task_metrics {
+                    tasks_rows.push(report::task_metrics_row(metric));
+                }
+
+                tasks_template = report::task_metrics_template(&tasks_rows.join("\n"));
+            } else {
+                tasks_template = "".to_string();
+            }
+
+            // Only build the tasks template if --no-task-metrics isn't enabled.
+            let errors_template: String;
+            if !self.metrics.errors.is_empty() {
+                let mut error_rows = Vec::new();
+                for error in self.metrics.errors.values() {
+                    error_rows.push(report::error_row(error));
+                }
+                errors_template = report::errors_template(&error_rows.join("\n"));
+            } else {
+                errors_template = "".to_string();
+            }
+
+            // Only build the status_code template if --status-codes is enabled.
+            let status_code_template: String;
+            if self.configuration.status_codes {
+                let mut status_code_metrics = Vec::new();
+                let mut aggregated_status_code_counts: HashMap<u16, usize> = HashMap::new();
+                for (request_key, request) in self.metrics.requests.iter().sorted() {
+                    let method = format!("{}", request.method);
+                    // The request_key is "{method} {name}", so by stripping the "{method} "
+                    // prefix we get the name.
+                    // @TODO: consider storing the name as a field in GooseRequestMetricAggregate.
+                    let name = request_key
+                        .strip_prefix(&format!("{} ", request.method))
+                        .unwrap()
+                        .to_string();
+
+                    // Build a list of status codes, and update the aggregate record.
+                    let codes = metrics::prepare_status_codes(
+                        &request.status_code_counts,
+                        &mut Some(&mut aggregated_status_code_counts),
+                    );
+
+                    // Add a row of data for the status code table.
+                    status_code_metrics.push(report::StatusCodeMetric {
+                        method,
+                        name,
+                        status_codes: codes,
+                    });
+                }
+
+                // Build a list of aggregate status codes.
+                let aggregated_codes =
+                    metrics::prepare_status_codes(&aggregated_status_code_counts, &mut None);
+
+                // Add a final row of aggregate data for the status code table.
+                status_code_metrics.push(report::StatusCodeMetric {
+                    method: "".to_string(),
+                    name: "Aggregated".to_string(),
+                    status_codes: aggregated_codes,
+                });
+
+                // Compile the status_code metrics rows.
+                let mut status_code_rows = Vec::new();
+                for metric in status_code_metrics {
+                    status_code_rows.push(report::status_code_metrics_row(metric));
+                }
+
+                // Compile the status_code metrics template.
+                status_code_template =
+                    report::status_code_metrics_template(&status_code_rows.join("\n"));
+            } else {
+                // If --status-codes is not enabled, return an empty template.
+                status_code_template = "".to_string();
+            }
+
+            // Compile the report template.
+            let report = report::build_report(
+                &start_time,
+                &end_time,
+                &host,
+                report::GooseReportTemplates {
+                    requests_template: &requests_rows.join("\n"),
+                    responses_template: &responses_rows.join("\n"),
+                    tasks_template: &tasks_template,
+                    status_codes_template: &status_code_template,
+                    errors_template: &errors_template,
+                },
+            );
+
+            // Write the report to file.
+            if let Err(e) = report_file.write(report.as_ref()).await {
+                return Err(GooseError::InvalidOption {
+                    option: "--report-file".to_string(),
+                    value: self.get_report_file_path().unwrap(),
+                    detail: format!("Failed to create report file: {}", e),
+                });
+            };
+            // Be sure the file flushes to disk.
+            report_file.flush().await?;
+
+            info!(
+                "wrote html report file to: {}",
+                self.get_report_file_path().unwrap()
+            );
+        }
+
+        Ok(())
+    }
+    */
 }
 
 /// Helper to calculate requests and fails per seconds.
@@ -2273,6 +2646,7 @@ mod test {
         assert!(!raw_request.update);
     }
 
+    /*
     #[test]
     fn goose_request() {
         let mut request = GooseRequestMetricAggregate::new("/", GooseMethod::Get, 0);
@@ -2478,4 +2852,5 @@ mod test {
         assert_eq!(request.total_response_time, 987657045);
         assert_eq!(request.response_time_counter, 8);
     }
+    */
 }
